backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.30
pydantic>=2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.1
httpx>=0.26.0
requests>=2.31.0
urllib3>=2.0.0
python-nmap==0.7.1
pytest>=7.4.4
pytest-asyncio>=0.23.3